Update myTFL_strings.py
dot function is associative, so I added the associativeTupl() method that can be used to check if two differently generated strings (with different repr) are actually the same string, for example:
dot(a, dot(b, dot(lambda, c))) and dot(dot(dot(a, lambda), b), c) both will result in the associative tuple of (a,b,c). I purposefully insist on not using any lists or strings!!! Just using functions and their arguments.